feat: implement role-based access control for member details viewing
- Remove ADMIN restriction from user gear endpoint (/users/:id/gear)
- Remove ADMIN restriction from user DKP endpoints (/dkp/users/:id/*)
- Allow PLAYER users to view other members' gear and DKP information
- Maintain authentication requirement for all endpoints
- Update endpoint comments to reflect new access permissions

Changes:
- users.routes.ts: Remove requireAdmin from GET /users/:id/gear
- dkp.routes.ts: Remove requireAdmin from GET /dkp/users/:id/history and /dkp/users/:id/summary
- All endpoints still require authentication (fastify.authenticate)
- ADMIN-only restrictions maintained for editing operations

This enables PLAYER users to view complete member information while
maintaining security for data modification operations.

+++ src/routes/raid-dropped-items/raid-dropped-items.schema.ts
@@ -54,8 +54,8 @@ export const updateRaidDroppedItemSchema = z.object({
 
 // Query parameters schema for raid dropped items
 export const getRaidDroppedItemsQuerySchema = z.object({
-  page: z.coerce.number().min(1).default(1),
-  limit: z.coerce.number().min(1).max(100).default(10),
+  page: z.coerce.number().int().min(1).default(1),
+  limit: z.coerce.number().int().min(1).max(100).default(10),
   raidInstanceId: z.string().cuid().optional(),
   category: itemCategorySchema.optional(),
   grade: itemGradeSchema.optional(),
